moving player

+++ src/main/java/org/amithalpert/Tools/GameObject.java
@@ -2,11 +2,11 @@ package org.amithalpert.Tools;
 
 public abstract class GameObject{
 
-    protected float x, y, velX, velY, speed;
-    protected float width, height;
+    protected int x, y, velX, velY, speed;
+    protected int width, height;
 
 
-    public GameObject(float width, float height, float x, float y) {
+    public GameObject(int width, int height, int x, int y) {
         this.x = x;
         this.y = y;
         this.width = width;
@@ -15,5 +15,6 @@ public abstract class GameObject{
         this.velY = 0;
         this.speed = 0;
     }
+
 
 }
